requirements.txt
qcloud-cos-v4
pyyaml